Add files via upload
This excel files contains code and example to latitude and longitude values from cells where gps location in ddmmss.ss is stored. The values in gps location column must be separated by -space- alone, and not contain the symbols for degree minute and second. The formula will match the location value entered in 'for analysis' sheet with the location value in the 'Raw' sheet and return the calculated lat lon value for that location.